third_party/blink/web_tests/platform/generic/mhtml/transfer_encoding_8bit-expected.txt

CONSOLE ERROR: Blocked script execution in 'transfer_encoding_8bit.mht' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
